This fork replaces lottie-ios 3 with lottie-ios `^2.5.0` and keep lottie-android the latest (at the time of forking).
The motivation of the hack is because of the bad performance of lottie-ios 3, which is the result of excessive usage of struct. For our case, the performance is so bad that when lottie animation is playing in the background, it almost freezes the app.

You can find out more about [the issue here](https://github.com/airbnb/lottie-ios/issues/895).